requirements.txt
selenium
pytest
PyFunctional
nose
lxml-3.6.4-cp35-cp35m-win32.whl
pytest-allure-adaptor